% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerizer-package.R
\docType{package}
\name{dockerizer-package}
\alias{dockerizer}
\alias{dockerizer-package}
\title{dockerizer: Create Dockerfiles For Your Analyses}
\description{
Create dockerfiles for your analyses
}
\author{
\strong{Maintainer}: Demetri Pananos \email{dpananos@gmail.com}

}
\keyword{internal}
