% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerize.R
\name{write_dockerfile}
\alias{write_dockerfile}
\title{Write a Dockerfile with appropriate system dependencies for your project}
\usage{
write_dockerfile(
  analysis_dir = ".",
  rocker_rstudio_tag = "latest",
  port = "8787",
  username = "rstudio",
  password = "rstudio",
  distribution = "ubuntu",
  release = "20.04",
  build = FALSE,
  container_name = NULL
)
}
\arguments{
\item{analysis_dir}{Directory to your analysis. Defaults to current directory.}

\item{rocker_rstudio_tag}{Tag for rocker/rstudio. See details.}

\item{port}{What port to open}

\item{username}{Username for rstudio}

\item{password}{Password for rstudio}

\item{distribution}{Which distribution of linux will your dockerfile run?}

\item{release}{Release number for distribution.}
}
\value{
Path to created dockerfile.
}
\description{
Write a Dockerfile with appropriate system dependencies for your project
}
\details{
Additional details...
}
